Upload/download scripts+gitignore
Made it ignore everything and force adding the needed files. This way it won't list a bunch of untracked files on each update.
Also removing all files in subfolders from the repo (before the new versions get added) so that if a config or a plugin(!) is removed it will remove it from the repo.