Updated the hunter job
Buffed market items, nerfed farmable items.

In other words, if you can sell the products of a mob in the market (nether stars, shulker shells, cows, and creepers), then the income is doubled or more. If you can farm the products of a mob in a mob farm (zombies, endermen, cows), then the income is halved or more.

I want to update the rest of the jobs with this basic premise with the intent to flood the market with goods that can be sold to other players. A boost to supply (by incentivising salable goods).