Added daily quests to the Builder job
We'll prototype this and see how it works. If this breaks, it's probably line 823 with two targets. I'm not sure if it lets you do that or not. I don't know if the pay command will work either. And I don't know if "daily" quests expire in 24 hours, or if they're just granted every 24 hours.